+++ NEWS.md
@@ -13,6 +13,7 @@ Red Panda C++ Version 1.0.4
   - enhancement: add Portugese translation
   - fix: crash when eval statements like "fsm::stack fsm;"
   - enhancement: add Traditional Chinese translation
+  - fix: index of the longest line not correctly updated ( which will cause selection errors)
 
 Red Panda C++ Version 1.0.3
   - fix: when oj problem grabbed by competitive companion received,